backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Django project managed by uv"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0",
    "django-environ>=0.11.2",
    "django-cors-headers>=4.3.1",
    "djangorestframework>=3.14.0",
    "psycopg2-binary>=2.9.9",
    "pillow>=10.2.0",
    "gunicorn>=21.2.0",
    "whitenoise>=6.6.0",
    "django-redis>=5.4.0",
    "requests>=2.31.0",
    "yt-dlp>=2024.3.10",
    "python-dotenv>=1.2.1",
]

[tool.uv]
package = false

[dependency-groups]

dev = [
    "pytest-django>=4.8.0",
    "black>=24.2.0",
    "isort>=5.13.2",
    "flake8>=7.0.0",
]
